Don't keep app bubbles on top
Previously the code was attempting to set app bubbles as always the
most recently accessed. This caused some weird ordering behavior.

Checked with UX and this should be changed to work exactly the same
as chat bubbles (i.e. whatever is most recently accessed - selected
and expanded, should remain on top).

Test: atest BubbleDataTest
Bug: 311345996

Change-Id: I9c70d4c470e27bbb5254a32007452c27c0f649d9