Add config_quickSettingsShowMediaPlayer
This CL adds config_quickSettingsShowMediaPlayer as a device resource
value to control the media controls support.
Currently, the state is controlled by the settings provider which we can
change at run-time so it’s useful for debugging.
But disabling the support by default (e.g., for ARC++) can be
challenging because we need to change the setting at runtime before
notification pipeline uses it, which results in a race condition.
Therefore, we need a compile-time value for it.

Note: I believe we can deprecate the settings provider one with this new
way. But in order to fix the ARC tests issues asap and to cherry-pick CL
to downstream branches easily, I would like to keep the deprecation task
out of scope for this CL.

Flag: NONE
Bug: 314876656
Bug: 192412820
Bug: 314889864
Bug: 314889882
Bug: 314886766
Test: m
Test: v2/arc-eng/tast_critical
Change-Id: Ic87edfadfa49d8d287f6514c554b8c881f8018f3