Do not autoscale textsize
Instead, use dp instead of sp. SP has been historically difficult for
the numpad keys as they do not fit in the buttons. These buttons should
not scale on scale density as they are not paragraph text. I looked at
iOS and the font on the pin pad also do not scale based on text scale.

Test: open user switcher, user edit info, keyboard, and then choose
avatar. This will open the bouncer and observe pin screen.
Bug: 301892345
Flag: NONE

Change-Id: I7633c4a8f9f728ad4136165e57c7f468dc0298f4