Extending systemUserOnly attribute to Providers and Services
systemUserOnly manifest attribute limits an android component to
be initialized and available only for the SYSTEM user. This is different from the singleUser attribute which instantiate the component on the SYSTEM user as a singleton and provides a proxy for other secondary users to access the SYSTEM user component.

Currently this attribute is only available for Activities and
Broadcast receivers, In the following change the attribute is extended to support Providers and Services as well.

Bug: 302354856
Test: atest PackageManagerShellCommandMultiUserTest#testInstallAppHavingSystemUserOnlyComponents -c
Test: manually tested by applying this attribute on a Service and a provider and making sure it isn't accessible and initialized when running on a Secondary user.

Change-Id: I3125bc0c336dfc11c40e327ce8992d5be3601922