Finish sync for activities under a drawn shared starting window
If a task contains multiple activities and the top one is translucent,
the task may not report sync finish even if the top activity is ready.
Because the activity behind is also visible, then the sync group still
needs to wait for it. But if the snapshot starting window has drawn,
the task should be covered by it. Then all visible activities belonging
to the task can be considered as ready to start transition.

This can also cover the case of hot launch of embedded activities.

Bug: 296817251
Bug: 317833037
Test: atest SyncEngineTests#testFinishSyncByStartingWindow
Test: Activity A starts translucent activity B in the same task
      and the same process. B will call sleep 1s in onResume.
      Return to Home and start the task of A. The transition
      should start in a short time by the starting window.
Change-Id: I3261f6b2137303ab3df16c836fc4c9dda6dfe338