Animation on fold for fold lock behavior setting
This animation is to be shown whenever the other screen stays turned ON
on fold.
The animation is linear light reveal of screen from left to right(when
device is in potrait mode) or top to bottom(when the device is in
landscape mode and hinge is at top when folded).
Test: FoldLightRevealOverlayAnimationTest
Bug: 290349959
Flag: ACONFIG fold_lock_setting_enabled TRUNKFOOD

Change-Id: Id050c34b3414360a2509bd4c1782c1f4327c235b