Remove an obsolete TODO comment
This is a follow up CL to my previous CL, which introduced

  InputMethodListBuilder

with a TODO comment to enrich API document on

  InputMethodSubtype#overridesImplicitlyEnabledSubtype().

The concept of explicitly-enabled subtypes and implicitly-enabled
subtypes are to now considered to be one of basic knowledge of Android
Input Methods. At least the following API documents assume the same
level of its understanding.

 * "allowsImplicitlyEnabledSubtypes" parameter in
   InputMethodManager#getEnabledInputMethodSubtypeList()
 * JavaDoc of
   InputMethodManager#setExplicitlyEnabledInputMethodSubtypes()
 * JavaDoc of
   InputMethodSubtype#overridesImplicitlyEnabledSubtype()

With above let's remove this stale TODO for now.

This is just a removal of a comment. There must be no observable
behavior change.

 [1]: I871ccda787eb0f1099ba3574356c1da4b33681f3
      302df9b492d430d4b8b0ac020237f7d292cfcaeb

Bug: 17347871
Bug: 18192576
Bug: 309837937
Test: presubmit
Change-Id: I24ca559f48ce9b56d7348d7727bb00a063f7b8ac