Remove android.graphics.RecordingCanvas.drawMesh(...) from current.txt
Previously, the `core/api/current.txt` erroneously and unnecessarily
listed `drawMesh(...)` as the only method in the
`android.graphics.RecordingCanvas` class. That was because there was an
bug in Metalava that meant it ignored a super method if that method
came from a hidden/inaccessible class. The accompanying change in this
topic fixes the bug and so it can be removed.

Listing `drawMesh(...)` method in `RecordingCanvas` was unnecessary
because `RecordingCanvas` extends `Canvas` which defines and provides a
concrete implement of `drawMesh(...)`. The `RecordingCanvas` does not
add any new methods, it simply overrides the methods from `Canvas` and
records their parameters.

Bug: 319826204
Test: m checkapi
Change-Id: I7e23b93659a87f13606ca2d6c4c3ceebee23901f