Add splitshade blueprint in compose.
Added the clock on one half of the screen and notifications on the other
side of the device.

I also added onUpdate on the clocks because when the current clock
changes, we need to replace the AndroidViews. This fixes an issue where
the time was not changing.

Bug: 316211368
Test: flexiglass on tablet
Flag: ACONFIG com.android.systemui.scene_container DEVELOPMENT

Change-Id: Ie89a81afa50eca13c0b150621ea6e278e350eadc