[ECM] Declare MANAGE_ENHANCED_CONFIRMATION_STATES permission
Declare the MANAGE_ENHANCED_CONFIRMATION_STATES permission in the core
manifest.

We recently defined this permission in this file, but, we didn't
declare it (with uses-permission). And, some system services technically depend on it (e.g., EnhancedConfirmationService calls SystemConfigManager calls SystemConfigService which enforces that the caller has this), so we should declare it.

(Note that this is strictly a hint. System uid has access to the
permission regardless of whether we add this.)

Test: presubmit
Bug: 320512579
Change-Id: I8736f3a86d93362d8fbc1696d518c3e635679368