Don't switch DND configs in ZenModeHelper upon reception of USER_UNLOCKED
This broadcast just means that the user's storage is accessible, not that it has become the "current" user.

The bug caused ZenModeHelper's configuration to point to the wrong user when secondary users are created and unlocked but never switched to (as some CTS tests do). In turn this caused the broadcasts expected by NotificationManagerZenTest to be sent to the wrong (and actually removed!) user.

Fixes: 322843652
Test: atest PackageInstallerArchiveMultiUserTest NotificationManagerZenTest
Change-Id: I962ab8f18573631fc040314fe186489b199fcb9e